New Commit 4/30/25
Some stuff I was working on during work. I fixed the buyer page so the buttons don't look so huge. Also, it is centered. All the background images should be centered now too. I'm working on the delete and update function logic right now but the buttons are on there for the interface.