Commit 5/6/25
The website is finished! (I hope)

Account implementation complete, now companies have only their own CRUD management of goods
Customers can view all products in the database from all stores in database. The buttons also show the respective categories
Update allows you to fix each tile
add product has autosends the category of product via drop down and whoever is logged on that is the name of the company that is sent (the area sourced for customer)

Customer feedback can be sent from customer side and companies read the messages sent to them. The message side has a drop down of all companies in the database to send to

MAKE SURE YOU PUT YOUR OWN DB FILE IN THE MAIN DIRECTORY FOR THE PATHING TO WORK!!!